properly enable OBDII enhanced data mode

+++ include/CAN/OBD2.h
@@ -82,7 +82,7 @@ void update_obd2_channels(CAN_msg *msg, OBD2Config *cfg);
  * @param value pointer to the value to populate, if the PID was found
  * @return true if the PID was found
  */
-bool OBD2_get_value_for_pid(uint8_t pid, float *value);
+bool OBD2_get_value_for_pid(uint16_t pid, float *value);
 
 /**
  * Sends an OBD2 PID request on the CAN bus.
@@ -90,7 +90,7 @@ bool OBD2_get_value_for_pid(uint8_t pid, float *value);
  * @param mode the OBD2 mode to request
  * @param timeout the timeout in ms for sending the OBD2 request
  */
-int OBD2_request_PID(uint8_t pid, uint8_t mode, size_t timeout_ms);
+int OBD2_request_PID(uint16_t pid, uint8_t mode, size_t timeout_ms);
 
 CPP_GUARD_END
